Revert "Consolidate AWQ quantization into single awq_marlin.py file"
This reverts commit 8cc700dd6a0d44c3b6983c0305af790b703f6b48.
